+++ src/DataBlockAggregator.h
@@ -34,6 +34,7 @@ using namespace AliceO2::Common;
 struct PartialSlice {
   uint8_t linkId;
   uint64_t tfId;
+  double lastUpdateTime; // timestamp of last block pushed
   DataSetReference currentDataSet;
 };
 
@@ -45,8 +46,10 @@ public:
   ~DataBlockSlicer();
 
   // append a new block to curent slice of corresponding link
+  // a timestamp may be given
   // returns the number of blocks in slice used
-  int appendBlock(DataBlockContainerReference const &block);
+  int appendBlock(DataBlockContainerReference const &block,
+                  double timestamp = 0);
 
   // get a slice, if available
   // if includeIncomplete is true, also retrieves current slice, even if
@@ -54,12 +57,19 @@ public:
   // iterated, returned in order of creation, older first
   DataSetReference getSlice(bool includeIncomplete = false);
 
+  // consider the slices which have not been updated since timestamp as complete
+  // they are flushed and moved to the "ready" slices
+  // returns the number of slices flushed
+  int completeSliceOnTimeout(double timestamp);
+
+  int slicerId = -1;
+
 private:
   /*
      uint64_t currentId; // common id of the blocks in current data set being
      built DataSetReference currentDataSet; // current data set being built
   */
-  const unsigned int maxLinks = 8192;      // maximum number of links
+  const unsigned int maxLinks = 32;        // maximum number of links
   std::vector<PartialSlice> partialSlices; // slices being built (one per link)
 
   std::queue<DataSetReference>
@@ -85,6 +95,10 @@ public:
   int disableSlicing =
       0; // when set, slicer is disabled, data is just passed through
 
+  double cfgSliceTimeout =
+      0; // when set, slices not updated after timeout (seconds)
+         // are considered completed and are flushed
+
   static Thread::CallbackResult threadCallback(void *arg);
 
   Thread::CallbackResult executeCallback();
@@ -100,6 +114,8 @@ private:
 
   std::unique_ptr<Thread> aggregateThread;
   AliceO2::Common::Timer incompletePendingTimer;
+  AliceO2::Common::Timer timeNow; // a time counter, used to timestamp slices
+
   int isIncompletePending;
 
   std::vector<DataBlockSlicer> slicers;