+++ src/receiverFMQ.cxx
@@ -32,6 +32,8 @@
 #include "RAWDataHeader.h"
 #include "SubTimeframe.h"
 
+#include "RdhUtils.h"
+
 // definition of a global for logging
 using namespace AliceO2::InfoLogger;
 InfoLogger theLog;
@@ -47,44 +49,6 @@ static void signalHandler(int) {
   }
   ShutdownRequest = 1;
 }
-
-// Implementation of a simple FMQ device
-class FMQReceiver : public FairMQDevice {
-public:
-  FMQReceiver() {}
-  ~FMQReceiver() {}
-
-protected:
-  int msgCount = 0;
-  int msgBytes = 0;
-
-  void Run() override {
-    while (!NewStatePending()) {
-      SetTransport("zeromq");
-      std::unique_ptr<FairMQMessage> msg(fTransportFactory->CreateMessage());
-
-      if (fChannels.at("data-in").at(0).Receive(msg) > 0) {
-        msgBytes += msg->GetSize();
-        msgCount++;
-        theLog.log("%d messages, %d bytes", msgCount, msgBytes);
-
-        /*
-        uint8_t *payload=(unsigned char*)(msg->GetData());
-        for (int k=0;k<16;k++) {
-          printf("%02X ",(unsigned int)(payload[k]));
-        }
-        printf("\n");
-        */
-
-        // cout << "Received message: \""
-        // << string(static_cast<char *>(msg->GetData()), msg->GetSize())
-        // << "\"" << endl;
-      } else {
-        usleep(200000);
-      }
-    }
-  }
-};
 
 // program main
 int main(int argc, const char **argv) {
@@ -147,8 +111,22 @@ int main(int argc, const char **argv) {
   } else if (cfgDecodingMode == "readout") {
     mode = decodingMode::readout;
   } else {
-    theLog.log("Wrong decoding mode set : %s", cfgDecodingMode.c_str());
+    theLog.log(InfoLogger::Severity::Error, "Wrong decoding mode set : %s",
+               cfgDecodingMode.c_str());
   }
+
+  // configuration parameter: | receiverFMQ | dumpRDH | int | 0 | When
+  // set, the RDH of data received are printed (needs decodingMode=readout).|
+  int cfgDumpRDH = 0;
+  cfg.getOptionalValue<int>(cfgEntryPoint + ".dumpRDH", cfgDumpRDH, 0);
+
+  // configuration parameter: | receiverFMQ | dumpTF | int | 0 | When
+  // set, a message is printed when a new timeframe is received.
+  // If the value is bigger than one, this specifies a periodic interval between
+  // TF print after the first one. (e.g. 100 would print TF 1, 100, 200, etc). |
+  int cfgDumpTF = 0;
+  cfg.getOptionalValue<int>(cfgEntryPoint + ".dumpTF", cfgDumpTF, 0);
+  theLog.log("dumpRDH = %d dumpTF = %d", cfgDumpRDH, cfgDumpTF);
 
   // create FMQ receiving channel
   theLog.log("Creating FMQ RX channel %s type %s @ %s", cfgChannelName.c_str(),
@@ -198,52 +176,80 @@ int main(int argc, const char **argv) {
         if (mode == decodingMode::readout) {
 
           int nPart = msgParts.Size();
-          // printf("msg: %d parts\n",nPart);
           if (nPart < 2) {
-            printf("Error : %d parts in message\n", nPart);
+            theLog.log(InfoLogger::Severity::Error,
+                       "Only %d parts in message, need at least 2", nPart);
             continue;
           }
 
           // expected format of received messages : (header + HB + HB ...)
 
           // first part is STF header
-          if (msgParts[0].GetSize() != sizeof(SubTimeframe)) {
-            printf("Error : header wrong size %d != %d\n",
-                   (int)msgParts[0].GetSize(), (int)sizeof(SubTimeframe));
-            continue;
-          }
-          SubTimeframe *stf = (SubTimeframe *)msgParts[0].GetData();
-          if (stf->numberOfHBF != nPart - 1) {
-            printf("Mismatch stf->numberOfHBF %d != %d message parts - 1\n",
-                   stf->numberOfHBF, nPart - 1);
-            continue;
-          }
 
-          // number of message parts matches number of HBFs in header ?
-          const int printStep = 100;
-          if ((stf->timeframeId == 1) || (stf->timeframeId % printStep == 0)) {
-            printf("Receiving TF %d link %d : %d HBf\n", (int)stf->timeframeId,
-                   (int)stf->linkId, (int)stf->numberOfHBF);
-          }
+          int i = 0;
+          for (auto const &mm : msgParts) {
+            if (i == 0) {
+              if (mm->GetSize() != sizeof(SubTimeframe)) {
+                theLog.log(InfoLogger::Severity::Error,
+                           "Header wrong size %d != %d\n", (int)mm->GetSize(),
+                           (int)sizeof(SubTimeframe));
+                break;
+              }
+              SubTimeframe *stf = (SubTimeframe *)mm->GetData();
+              if (stf->numberOfHBF != nPart - 1) {
+                theLog.log(
+                    InfoLogger::Severity::Error,
+                    "Mismatch stf->numberOfHBF %d != %d message parts - 1\n",
+                    stf->numberOfHBF, nPart - 1);
+                break;
+              }
+              // number of message parts matches number of HBFs in header ?
+              if (cfgDumpTF) {
+                if ((stf->timeframeId == 1) ||
+                    (stf->timeframeId % cfgDumpTF == 0)) {
+                  printf("Receiving TF %d link %d : %d HBf\n",
+                         (int)stf->timeframeId, (int)stf->linkId,
+                         (int)stf->numberOfHBF);
+                }
+              }
+            } else {
+              size_t dataSize = mm->GetSize();
+              void *data = mm->GetData();
+              std::string errorDescription;
+              for (size_t pageOffset = 0; pageOffset < dataSize;) {
+                if (pageOffset + sizeof(o2::Header::RAWDataHeader) > dataSize) {
+                  theLog.log("part %d offset 0x%08lX: not enough space for RDH",
+                             i, pageOffset);
+                  break;
+                }
+                RdhHandle h(((uint8_t *)data) + pageOffset);
 
-          for (int i = 1; i < nPart; i++) {
-            if (msgParts[i].GetSize() % 8192) {
-              printf("size not matching expected HBF size (multiple of 8k)\n");
-              break;
+                if (cfgDumpRDH) {
+                  h.dumpRdh(pageOffset, 1);
+                }
+
+                int nErr = h.validateRdh(errorDescription);
+                if (nErr) {
+                  if (!cfgDumpRDH) {
+                    // dump RDH if not done already
+                    h.dumpRdh(pageOffset, 1);
+                  }
+                  theLog.log(InfoLogger::Severity::Error,
+                             "part %d offset 0x%08lX : %s", i, pageOffset,
+                             errorDescription.c_str());
+                  errorDescription.clear();
+                  break;
+                }
+
+                // go to next RDH
+                uint16_t offsetNextPacket = h.getOffsetNextPacket();
+                if (offsetNextPacket == 0) {
+                  break;
+                }
+                pageOffset += offsetNextPacket;
+              }
             }
-
-            /*// iterate from RDH to RDH in received message
-            int nBlocks=0;
-            char *payload=(char *)msgParts[i].GetData();
-            for (unsigned int offset=0;offset<msgParts[i].GetSize();){
-              o2::Header::RAWDataHeader *rdh=(o2::Header::RAWDataHeader
-            *)&(payload[offset]);
-              //dumpRDH(rdh);
-              offset+=rdh->blockLength;
-              nBlocks++;
-            }
-            //printf("%d CRU blocks in HBF\n",nBlocks);
-            */
+            i++;
           }
         }
       }
@@ -260,7 +266,6 @@ int main(int argc, const char **argv) {
       }
     }
     // printf("releasing msg %p\n",msg->GetData());
-
     // std::cout << " received message of size " << msg->GetSize() << std::endl;
     // // access data via inputMsg->GetData()
 
@@ -286,7 +291,7 @@ int main(int argc, const char **argv) {
   // other implementation with a full FMQ device
 
   std::vector<FairMQChannel> channels(1);
-  FMQReceiver fd;
+  FairMQDevice fd;
 
   // todo: check why this type is not public in FMQ interface?
   typedef std::unordered_map<std::string, std::vector<FairMQChannel>> FairMQMap;
@@ -297,7 +302,7 @@ int main(int argc, const char **argv) {
   channels[0].UpdateAddress("tcp://localhost:5555");
   channels[0].UpdateRateLogging(0);
   channels[0].UpdateSndBufSize(10);
-  if (!channels[0].ValidateChannel()) {
+  if (!channels[0].Validate()) {
     throw "ConsumerFMQ: channel validation failed";
   }